dapp: adopt fastlane to make internal test drafts
At the moment, Fastlane can only finalize its task when uploading a
draft to the internal test track of the Google Play store. Else it will
fail with an error of the related API. This appears to be related to the
fact that our application is not yet fully configured/set-up in the Play
Store. Though the documentation is really confusing about this part. To
still make progress here, we limit it to this scope for the moment.
Later we need to extend the lanes to include production releases anyway.

Draft releases can be finalize in the Web-UI of the Google Play Console.

This commit includes some formatting as recommended by the language
server (in first place single quotes for double ones).