dapp: fix issue with blocked hubs for transfers
If a user has a channel with another node and this channel is not open
(anymore), he can't do transfers to it anymore, though he could
(eventually) do a mediated transfer to this node.
This is the case because the address input field for transfers filters
direct cannel partners which are not open anymore and prints an error
for them. This relates to the old days when the LC was meant to be used
in first place with hubs only. This has changed. Therefore this
behaviours is unwanted now.