sdk: add dispatchRequestAndGetResponse custom rx operator
This operator faciliates the pattern where one needs to dispatch a given
request (async action) to the output of an epic/observable, then wait
for the response (success/failure) and proceed with the "normal"
observable. This "shortcutting" the request to the output is performed
with a dispatch parameter on the project function.