next: sdk: fetch presence from PFS
This removes completely the requirement to having access to peer's
presence directly from the matrix server, and instead replace it with
the requirement for the PFS to know (already required for mediated
transfers) and provide that presence (implemented in `next` RSB).
Instead of searchUserDirectory, we request the PFS
`/address/.../metadata` endpoint for its view of latest peer's presence.