dapp: fix header line break on mobile screens
On mobile screens with a very narrow display, the title in the header of
the dApp starts to insert line breaks when the text is too long. This
does not looks good.
As solution the whole header tries to narrow down to safe space in the
width. Additionally the font size shrinks a little to make it look more
appealing. Finally a new arrangement of the components in the header
with a grid got established to have better control of the centered titel
and weight distribution depending on if a back button and or identicon
exists.
This got visually tested on a width of 320px and all sub-pages (except
the one to finally open a new channel).