dapp: split web3 provider class for complexity
The initiator for this change was the CodeClimate tool. It detected
that function to determine which provider to use and to create the
according one got too complex. Therefore the higher function now only
determines which provider to create and then calls individual functions
to create those providers. Additionally the type declaration for the
provider instance got added and made more restrictive.

This code splitting is also valuable for the upcoming connection
manager. Though the tests for this class were and still are not really
exhaustive. With a future more generic interface this should improve.