dapp: add connection pending dialog
Sometimes the connection process get stuck because the user does not
confirm the wallet connection or signing the subkey. Sometimes this is
caused because the Wallet does not pop-up or similar. In any case, if
the connection takes a while we want to inform the user that he might
need to take action to continue here.
Therefore the connection pending dialog got introduced. It gets shown a
couple of seconds after the user clicked the connect button. Therefore
it uses the loading state of the store. That one got adopted to be only
true while the app is really loading. Before it was per default always
loading until being connected.