dapp: add reset connection button
When a the establishment of the connection got stuck, the last resort is
to reset the connection. This is especially the case for the
WalletConnect provider as here the local session state mut be cleared.
Else after the reload of the dApp it would pick up the old session
again.