sdk: implement new webRTC signaling protocol
This new signaling is backwards compatible with the previous one,
because it simultaneously listens as callee and also calls as caller,
and races the first channel which can get opened. It also replaces the
infinite caller retries by retrying a few times only when certain events
happen, e.g. messageSend.request