dapp: remove connect dialog and first time connect
The connect dialog does nothing anymore. It shows the first time the
user connects and has the configuration option to use a subkey (which
defaults to true). Besides that is just delays the connection. Also it
does not add any valid information in terms of that the subkey gets
regenerated each time the user connects anyways.
But after all the major reason to get rid of this stuff is because of
the upcoming connection manager. Thereby it will become obsolete anyway.
With this step to remove it, it helps to clean the code a little to
provide a more flat basement to draft on.