dapp: simplify connection status management
Due to multiple historic changes, the connection status got kinda
complicated. This could be simplified by getting rid of the global
loading state. Also is the status now directly part of the state and is
no more a derived property. This simplified a few old test which haven't
been refactored yet. In general it simplifies the usage/management and
readability.