dapp: add direct error handling for raiden service
The too complex RaidenService connect function gets more and more
simplified and less tight to its environment. One part it still did was
to evaluate different errors of the SDK and push error states to the
Vuex store. There they got picked up by the Home view component which
uses another component to translate and display them to the user.
Futhermore the stores state got used for the condiational flow.
This is completely over the top. The Home view component calls the
RaidenService plugin directly. Thereby it can get errors directly and
handle them directly. The whole detour is just not necessary. Moving the
error handling directly into the caller is more naturual and heavily
reduces the complexity of the whole information flow and the
RaidenService class.

Furthermore the translation of errors was kinda cumbersome as well. Plus
is was hard to extend, especially for a more generic scenario not only
for connecting. Therefore this got simplified as well, allwing to strip
the whole NoAccessMessage component. In future extension it probably
makes sense to have a small Vue plugin or a Mixin to map the error
codes to translation keys at a single place.