dapp: fix backup state upload
The dialog which handles the upload of a backup state got broken. This
was causes by the recent changes of the RaidenService connect function.
But actually the dialog must not call the connect function anyway.
Rather the backup is stored into the Vuex store and there gets picked up
by the next connection procedure. This was rather an uncontrolled
behavior which eventually could cause bugs. Cleaning this up also
removes the origin issue.

Furthermore while working on this, I detected that the backup state
never gets cleared from the store while the browser tab is active. This
means is the user connects with a backup, then logs out and reconnets
again, it would again use the backup state and overwrite any change that
got applied in-between or causes a failure at the SDK. To fix this, the
backup state gets cleared when the connection prodedure finishes.