dapp: refactor web3 provider with new design
This commit does not add any new feature to the dApp, but refactors the
design how the dApp gets an Ethereum connection and account to
instantiate the SDK. This new approach has been developed to prepare the

The new design introduces an abstract class which represents an Ethereum
connection, holding the actual network/RPC provider and the account to
use for signing. This abstract class gets them implemented by all the
various connection methods the dApp provides. The intention is to have a
nice and clean type system where the logic to instantiate the SDK and
connect the dApp does not care about where the connection comes from,
but just takes everything that implements this interface. Thereby the
logic handling the connection process can be easily decoupled and stays
static, no matter how many new connection methods get added or removed.
Furthermore is must not care (not yet ready) about how the connections
get configured. This allows the user to take interactively control over
the configuration instead of a static file served by the web-server.

After all, for the moment, the Home view component that handles the
connection just gets a connection instance and passes its properties to
the SDK. The selection and instantiation of the connection however is
now done in the former components was well. This is the part that
becomes replaced with the connection manager at a later point.

Furthermore the does this approach improves the testability of the
touched components. The connection process module is static and
independently testable from the connection method. And each connection
method can be tested individually. Mocking becomes (theoretically) easy.
All by the bounding interface that tights everything together.