dapp: add availability field to connections
Ethereum connections can be available or not. For the moment is
especially the case for the injected provider connection. This adds that
field to all kind of connections. This allows to visually inform the
user that a connection is not available or just hide it.
The field can be statically controlled (e.g. via a configuration option)
or dynamically evaluated (e.g. for the injected provider).