dapp: export user settings to own store module
The user settings will become extended. For example it is planned to put
the ethereum connection options there. This will require some additional
logic and will grow over time. Therefore it make sense to export it into
its own store module. Thereby we don't clutter the main store and the
module can be a focused selection of storage, individually tested and
maintained.