dapp: add connection options to vuex user settings
This extension of the user settings allows to make the Ethereum
connection options as provided by the user can be easily preserved and
reloaded.

This is a preparation for the connection management by the user. For
this commit the connection options are still derived from the
configuration file. Later this will allows to save the users connection
options and pre-fill the connection dialog, allow for auto-logins,
change the options somewhere else and more.