dapp: improve action buttons loading state
The ActionButton component had a boolean property called `loading` that
sets it into a loading state. Furthermore it had another boolean named
`syncing` to tell the component when in loading state, show an indicator
and a different text. The used text was using automatically a
translation from the Home view component.
To fix this awkward behavior, a parents component now provides a
property called `loadingText` (note that the "normal" text property is
called `text`). If this exists and state switches to loading it shows
this text next to the loading indicator.