dapp: introduce connection manager component
All the previous refactorings in regards of the connection establishment
have worked towards this step. Everything about the initial connection
is handled by this component. The Home view component is now only about
the view and includes this component.
After all this was basically just a split of the large Home view
component into two with better separation of concerns. Thereby it is
mostly just code that has been moved without any new feature. But some
naming schemes in the new ConnectionManager component could have been
simplified/shorten as of the now clear context.

Note that the ConnectionManager does not care about what happens once
the connection got established. Rather the Home view component as part
of the higher level routing is observing the connection state and
automatically forwards the user if a connection got established.
To mention is also that the unit testability has incresed as well

In the next steps this new component will change its implementation to
give the provider option determination into the hands of the user and
get rid of the configuration file (at this part of the project).