dapp: use dialog to establish link for connection
This makes the WalletConnectProvider dialog linking the EthereumProvider
and emit it to the ConnectionManger. This finally removes the automated
provider detection via configuration and put it into the hands of the
user. Though this temporally removes the possibility to connect via
injected provider or direct RPC connection. They will get added later.

This includes some styling changes which might look off for this commit.
They have been added to make sure the dialog can fit on the agreed
smallest screen size.

This will temporally make the end-to-end tests fail.