dapp: fix e2e tests for connection manager
The ConnectionManager now requires the user to manually link the
Ethereum provider. The end-to-end tests worked so far via an automated
connection using the configuration file. Additionally it was just
clicking the single connect button.
As this whole process does not work
like that anymore, but the end-to-end tests are based on using a direct
RPC provider, they got temporally fixed by using an automated connection
if the configuration exists.
As soon as the ConnectManager got extended to also include the
DirectRpcProvider, this will change and the configuration file for the
end-to-end tests can be ditched.