dapp: clean up wallet connect dialog component
During the work of additional features it has shown that the structure
if the component and how it handles the provider options is hard to use.
Therefore the component got restructured. The major change is that
each option has its own input and only the visibilty and usage gets
changed. This (will) allow(s) a better usage when it comes to integrate
the Vuex store into this component.
Furtheremore the styling and alignment could have been strongly
simplified. The only remaining "ugly" thing is the manual styling of
the text input fields.
Finally the strong separation of all provider inputs lead to a more
linear naming for classes, translations and component state.