dapp: export shared provider dialog logic to mixin
All dialogs to configure the options of an Ethereum provider and link
to it share the same logic. To reduce duplicated code and simplify those
dialog components, this shared logic got exported into a Vue mixin
component. This also reduces some of the duplicated test cases between
the already existing dialogs.
This will heavily simplify the creation of new provider dialogs.