dapp: adopt e2e test to use provider dialog
So far the end-to-end tests where using a workaround/fix to use the new
connection manager. This got not finally replaced with a proper workflow
as it should be. Therefore the tests no use the recently added direct
RPC provider its dialog, insert the options and connect.

This change makes the configuration file for the end-to-end tests
obsolete.