dapp: add a description to provider dialogs
As the configuration of Ethereum providers happens inside dialogs, there
is no option to show one of the info pop-ups. To still provide a little
information to the user, each providers dialog can now include a short
description at the top. This should give the user a little more insight
what stands behind those broad titles.
Furthermore it allows integrate messages that an option is not recommended.
Finally the injected provider which has no options to configure does not
look so wrongly empty now.