dapp: fix moved use raiden account setting
The root state of the store formerly included the property if the user wants
to use a Raiden account or not. This was moved into the new stores
submodule for the user settings. Unfortunately it was forgotten to
adapt this usage in some components. This commit is fixing that.

This is another indicator that our linting/type safety for pure string
in regards of the Vuex store does not work. We need to find a solution
for that.