dapp: make route guards purely functional
This improves the unit testability and independence of the route guard
implementations. In first place this removes the dependency to the store
module by making it a function parameter. In case of the global guard
function it has become part of the context it gets bound to.
This allows to have clean tests and full control over the store with
clean states between the cases.