dapp: add new route when no token is connected
So far the transfer route was doing two jobs at once. In case the user
connects the first time and has no token with a channel, it displays a
special component which guides the user to connect his first token. Else
it shows all the transfer components. Besides this is not actually nice,
it also lead to the wrong route title `Transfer` with the respective
info panel, even this is not correct.
To not just fix the symptoms but the issues cause, this commit splits
the transfer route into two. This leads into a simplified/straight
forward transfer route. And a second route which guides the user to
connect his first token. The latter is basically the former child
component which became a full route component now.
The decision which of the routes to show is handled by a route guard. If
the user has no token with a channel it will show the new route. Else -
the common case - it shows the transfer route. Exceptions are the
account routes as for the connection guard and any route that is
necessary for opening a channel. If the user successfully went through
the process of opening a new channel, the final navigation will now work
as the guard has nothing to complain about anymore.