dapp: fix connetion router guard for account
The connection router guards needs to allow the access to some account
routes to allow the user to change the settgins or do backups.
Unfortunately it was actually allowing the access to all account related
routes, including those which need a connection. This should not be
allowed. Therefore this got fixed.