dapp: fix flaky end-to-end tests
The end-to-end tests of the dApp have started to fail sometimes. After
long research it turned out that this is a timing issue. The issue that
bubbles up as a result was that certain sticky buttons where hidden by a
notification snackbar. Therefore they could not be clicked and tests
failed.
The snackbar that was hiding the button was the backup reminder. It gets
displayed for 60s when the user does not dismiss it manually. Also
removing the notification from the panel has no effect on it. On older
versions, the tests apparently took a little longer. At this time in the
scenario the snackbar already disappeared when one of these button
should be clicked.
Now this was apparently a  game of a view (milli)seconds where the
snackbar either disappeared too late or early enough.

The solution for this issue is to make the backup reminder notification
to never disappear automatically and dismiss the snackbar in the tests.
It must not disappear automatically, because else it would remain the
uncertainty if the snackbar is still there when it get attempted to be
closed. Though there could have been established a conditional
interaction, having a sticky snackbar for this notification is fine.
Doing a backup is important and the user should dismiss it himself.
That's a common behavior among many dApps.