dapp: declare custom vuetify styling variables
This SCSS file gets automatically sourced by the Vuetify loader and
overwrites the default of the package.
The customized variables target the dialog component. It reduces the
paddings and margins to improve on small screens by having more space
for the content. It is not possible to define variables conditionally
for small devices. But these values look also very good on large
screens.
This is a part of the fix for many broken dialogs on small screens.