dapp: simplify and improve action button component
The ActionButton component is a mess. And it is the cause for many
broken UI layouts on smaller screens. The root of the issue with this
component is that it is not a pure button but is actually surrounded by
a row and column plus more. This is not necessary and hurts more than it
helps. If it ever have helped at all.
Furthermore many parts could have been simplified, totally removed (dead
code/unused) and more. After all this component is now way more straight
forward and very flexible to use.

This change required many unit tests to adopt. The tests of other
components where making assumptions about the implementation of the
ActionButton component. This should not be the case. Therefore they got
improved by acting on the component itself. The Vue dev-tools provide
that functionality exactly for that case.