dapp: simplify all dialog components
The dialog components break on small screens. This got partially solved
by providing more space via customized Vuetify variables. Furthermore it
got improved by making the ActionButton component more flexible.
But after all, each dialog component was a mess. While it is not target
of this commit to make them all perfect, they got all stripped down a
lot. They were abusing the usage of Vuetify rows and columns with cards.
The result are tons of useless lines in the template. They clutter the
code, making it very bad to read what a component intends to do. But
finally all these Vuetify components come with lots of styling,
especially paddings and margins which make them very bad to use on small
screens. The basic card component in a dialog is very powerful and using
plain flex styling (as it is per default via Vuetify) is perfectly fine.

After all dialogs got touched here, they should be refactored at some
point in time. 70% of their code is shared or basically the same. Either
copy-paste or just different naming, or optional stuff. To maintain this
gigantic amount of dialogs which all are just 2-3 different kind of
dialogs is a huge amount of not necessary work.