dapp: use more sticky buttons for routes
The routes the user visits one after each other to do a transfer use
sticky actions buttons. This has the advantage that they don't jump
around from route to route and look just great.
For the pretty similar flow of opening a channel, where the user also
visits many routes in a row, this was not the case. Making them sticky
here as well provides a more uniform look and also prevents the issues
described above.

Furthermore the spacing on some of the screens had been improved to make
it work better on small screens and look more clean/uniform.