dapp: fix transaction history on small screens
This converts the former grid layout into a flex layout. This allows to
use the space more efficient and looks better. It results also into a
way less verbose template, making it easier to grasp what the component
does.
To save necessary space, the text `Sent to` and `Received from` got
removed. The icon and the amount sign should be enough to display this
information.