dapp: add sign prop to amount display component
This simplifies the usage of a sign before the amount to display. Else
it will be always two components which are hard to align and can get
broken. Putting it into the display component itself makes it more
stable and easier to use.
Furthermore the translation for the sign has been removed. This is a
static sign and will never change with the language. This is just
"math".