next: sdk: receive address_metadata and use it for transfer.request
Services now return a new `address_metadata` field which maps addresses
to their respective metadata/presence object. We need to handle this,
and pass it in a transfer's metadata. Besides this, mediated transfers
should try to use this mapping in the incoming transfer's metadata to
optimize not having to request metadata from service to send the output
message. This usage, despite validated by signature, should be
constrained to this transfer only (for now).