dapp: make sure notification queue is sorted
The notification queue which is used to display snackbars to the user
should be sorted. Else, if this property would not be guaranteed, it
could be that the user sees the notifications in a wrong order. The
list used for the notification panel was already sorted. Therefore it
was simple to share this feature between both store getters.