dapp: fix notification snackbar reactivity
The snackbar to show notifications to the user was not 100% working correctly.
The major issue was that if the user removes a notification from the
notification panel or even clears all notifications, this was not recognized by
the snackbar. A resulting issue of this problem was that it was not possible to
close a notification snackbar if the related notification got removed from the
store. This was especially bad if the notification had no timer to
automatically hide the snackbar.
Furthermore the implementation of the snackbar component was kinda confusing
and not obvious to understand. Even with the knowledge what it intends to do.
To fix the component, it got drastically simplified. Therefore it just displays
the top (oldest) entry of the notification queue from the store. When the
notification timer runs out or the user dismisses it, the notification gets
marked as shown. No more and no less. As result with the reactivity of the Vuex
store, this will change the notification queue, which turn will change the
getter of the component which links to the top entry of this queue. If the
queue is empty, nothing gets shown.

Finally the template could get simplified a little. This is a tiny component
with a little logic and a little to display. This is now better reflected by
the code.