dapp: simplify e2e tests for notifications
The previously added extension to the end-to-end tests of the dApp is not
necessary anymore. Its task was to dismiss a notification snackbar. This was
necessary to close the sticky backup notification which hides/blocks input
components for the continuing test run. This was necessary due to a bug in the
notification snackbar component. Now, as that one got resolved, the
notification gets hidden by the removal of the notification from the panel.
This interaction was already there before. The combination of both interactions
is actually not possible. Therefore the second one gets removed.