dapp: fix token overlay for small screens
This works again because it simplifies the template code. It removes the
not necessary usage of Vuetify row-, and col-components which add a lot
of margins and paddings and grids which can brake. The optimal solutions
is to not abuse the list item component, but use it properly. This even
fixes the ancient issue that the connect new token list entry was out of
wrongly aligned.