dapp: fix channel list on small screens
On small screens, the withdraw button for a channel disappears due to
missing space. This issue got fixed by three improvements. The first and
major issue is the removal of the state information for a channel.
Channels are grouped by their state with big headings. Therefore it is
just not necessary to show this information again. This is a similar
decision as recently for the transaction history list items. This change
allowed to remove a lot of no more necessary code and simplification.
The second change was the reduction of the space between the buttons on
mobile screens to make the content more narrow.
Finally the third change is to adopt the Vuetify SCSS variable for the list
item padding to give more space again.