dapp: fix notification cards on small screens
The composition of the notification cards content causes the issue that
it does not flow well. Thereby it leads to an awkward result for small
screens. This could be solved with proper flow styling. The white-space
property is used instead of a static padding to work better on any
line-break scenario.

Furthermore the component got slightly improved. For example could the
watch, plus internal state be replaced with a straight forward getter.