dapp: allow providers to be disabled by config
The administrator of a dApp server should be in charge to disable specific
ethereum providers. So do we aim for disabling the direct RPC provider in our
publicly hosted version due to security issues. In contrast to not available
providers they should be completely hidden in the UI.
To achieve this, the configuration files got extended to include such data.
Then the EthereumProvider base class implements a function to check if the
current provider is listed or not. As result providers are enabled per default.
Note that the availability of a provider has to be distinguished from this. The
availability refers to technical inspection of the environment the dApp runs
inside.
It is not really possible to integrate the disabled property of a provider into
the constructor to make it fail like for the availability. The root cause for
this is the asynchronous reading of the configuration file. Though as the
disabled property is not a technical limitation this should be fine. It is
intended to be used to conditionally render providers in the UI to give the
user access to it or not.