dapp: adapt connection manager to be more generic
The connection manager component grows a lot for every newly added provider and
every related feature that gets added. This clutters the component quickly and
makes it hard to change it.
Therefore the component was made a little more generic. Instead of adding
explicit state, getters and methods for every single provider, it is using now
a list of providers. All states and methods are not based on the name of each
provider. Also the template now renders the buttons and dialog with a loop,
keeping it very organized.
While not optimal, this makes the component less verbose and more clean. Adding
or removing providers is now a single change in the provider list. Everything
is generic enough.