dapp: fix empty logs in production mode
When the dApp runs in development mode, the whole logging works fine. The
loglevel libraries factory method could be extended and pipe log messages into
a database. From there the user can download a dump of the database.
In production mode this was not working and the download result into an empty file.

The origin of this issue are the code optimization tools which are part of
the buid-chain for dApps production mode. One feature of them is to remove all
functions `.name` properties to save a lot of bytes in the final bundle.
Unfortunately this property is used in the SDK to filter logging factory
methods. The origin for this filter is to prevent Matrix monkeypatching to
overwrite the logging factory.
To solve this issue, the filter must not rely on the methods name. Therefore a
custom property gets attached to the method. As nobody else than us does this,
this works. Its just an agreement with ourselves as it was before with the
methods name.

Furthermore it was necessary to give `loglevel` its default factory method an
empty context object. Else it runs into some undefined context errors.