dapp: fix duplicated navigation error
When the user was doing a backup, after he landed on the home route again, an
error gets thrown because of duplicated navigation. The cause for that is the
following. When the raiden service gets called to dump the database of the SDK,
it also stops the SDK. If the SDK gets stopped, an event hook automatically
redirects the user to the home route. But then the backup dialog manually
attempts to navigate to home again.
Fixing this bug is easy as it just requires to remove the manual navigation and
use better function naming.

Apparently this solves the end-to-end tests as well.