chore: add github action to block fixup commits
Fixup commits are very useful tool when working on bigger features or react to
feedback of a PR review. But before the PR gets merged they must be squashed
into their referring commit. Sometimes this get forgotten and a fixup commits
makes it into the master branch. This is not nice and should be avoided.
Therefore this adds a new action to the GitHub worflow checks of a PR to verify
that there are no fixup commits. If the action fails, the merging is blocked
and the author must first resolve it.