dapp: make unit tests fail on console errors
The output for the unit tests should be short and clean if they pass.
Unfortunately this is not the case if some specifications do not mock
correctly. Though the test cases are passing, its not 100% correct. But the
worst of all issues is that this clutters the test output. Developers get used
to tons of output with many warnings and errors. This should not be the case.
To prevent this once and for all, this additions makes test fail if the console
warning or error channel gets used. This forces the developer to fix the
mocking (or what ever causes it) to get passing tests.