sdk: adjust PFSFeeUpdate timestamp's & imbalance rlp encoding
This message was not 100% complying with what's emitted and expected by
PC and PFS. Our generated timestamp was a string at the second
resolution, which was accepted fine but differred from PC's and
prevented us from being able to verify a signature from them. Also, the
`imbalance_penalty` array had wrong RLP encoding for integer zero, which
prevented proper validation on PFS if this field was defined and
contained zeroes (usually the case). We hadn't noticed because the tests
didn't have this field defined and PFS always accepted our messages
because they were `null` so far.