cli: extend ci setup to release on npm registry
The raiden-cli package should be released on the NPM registry now as well.
Therefore the CircleCI job responsible for the publishing got generalized as
command for any workspace in the repository. The former job now calls the
command to make releases for all workspaces we intend to release.
As extension this command is also able to optionally upload a package as
tarball. This is necessary for the CLI as a bundle.

The bumping of the package version has always included the CLI already.
Therefore no further changes are required.