cli: apply default fees for incomplete tokens
The functionality of how the clients handle the default/missing fee
values for a token works differently in the Python Client than in the
Light client. Therefore, if the user defines any fee value (e.g. flat
ones only) for a token via command line arguments, we must ensure that
all missing fee model values for this token get set to the default. Else
these would be zero which is not how the Python Client works.