dapp: fix service workers for capacitor build
The Capacitor environment does not include a service worker. Therefore it fails
to be loaded with the necessary feature checks enabled (where disabled for test
purpose). Furthermore the Capacitor build is in production mode. Which means it
intends to start the service worker.
To solve both issues, a new environment variable has been introduced to allow
disabling the service worker feature manually. This variable gets also used in
the very basic index.html file to check the necessary features. Using this
variable makes the service worker check conditionally. Finally the servicer
worker assistant plugin only starts the service worker if in production mode
**and** if it is enabled.
To make the index.html rendering always working, there must be a default value
for this variable. Else this would be just empty, which would make it invalid
JavaScript code.