sdk: don't send parallel withdraw requests
While testing the previous commit, I got a LC node out-of-sync with a PC
because PC rejects WithdrawRequests while there's still a pending one.
Technically (per protocol), there's nothing preventing parallel
WithdrawRequests. They're idempotent on the amount, and the sender can
decide which/when to send any of them. We've implemented supporting
parallel withdraw requests, but since we want to comply with PC, we'll
assert and ignore user's new requests while there's any pending.
Additionally, as a special case, if the pending request is already
confirmed by partner, we resend it to retry sending tx, in case the
previous request failed.