cli: put package into the @raiden_network namespace
Unfortunately the package name `raiden-cli` at the NPM registry is already
blocked and the conversation with the owner did not happen so far. Therefore, as
an alternative, we will publish the CLI under the `@raiden_network` namespace.
Though we try too keep the adaption in the repository as minimal as possible,
sticking to the simple `raiden-cli` name where ever possible (e.g. for the
directory).