cli: publish package without tarball
Instead of manually construct a tarball to release, specify within the package
description what belongs into the tarball Yarn builds itself. This also
simplifies the automated release in the CI pipeline.

Furthermore the binary entry of the package has been named explicitly named to
avoid any issues/uncertainties because of the namespace.

Finally the package description file got re-ordered a little. The intention was
to group the properties more by purpose/usage. This means that first comes
everything descriptive that is especially parsed by package registries. Then
come the scripts to operate on the package, then how to build the package with
the entry points. At the ends come the dependencies of the package.
This bloats the commit a little, but it should be weighted out by the better
order.