chore: add new github action to check licenses
The so far used license checker has run within the continues integration
pipeline. Unfortunately does it not support the Yarn workspaces. As a result it
never reports, nevertheless if there are not allowed licenses used or not.
Unfortunately it looks like there is no license checker out there which fits our
needs. These are:
  - work with Yarn workspaces
  - only check first level production dependencies
  - do not match licenses by regular expressions
  - run fast and in parallel

This lead to the here added custom implementation of a license checker. It is
intended to be used with GitHub actions. It only checks a single Yarn workspace
at a time. Though it is meant to be run in parallel for all workspaces.
The application of the action has to to follow.