sdk: fix outdated unconfirmed udcDeposit.success actions
When depositing to UDC, after the tx just got mined, it emits a
udcDeposit.success. So far, it was being emitted with the value it just
got from a new effectiveBalance call, which may be outdated because the
tx didn't get confirmed yet. Instead, we now calculate the actual "final"
effectiveBalance to be used there since we know it from previous
balance.