dapp: remove idb package usage for service worker
While this package makes it super easy to work promise based with the
indexed database API, it is super hard to write deep tests with it. As
for the service worker will aim to be partially
integration tests, it is not just fine to mock the whole package as is.
I tried to mock the API and still use `idb`. But is was a lot of work
and in the end it was still not fully working. The package
implementation is quite complex. Finally it was so much more simple to
just use the API directly and ditch the package. This allows made the
mocking very easy as only necessary parts have to be provided.