dapp: add unit tests for service worker index
This are finally the tests for the whole service worker as is. As they
also depend fully on the functionality of the Workbox package, they can
be partially considered as integration tests. As we will have only
end-to-end tests on top of that, this is really good.
It includes mocks for all the missing parts of a service worker
environment.

It also includes a little change to the service worker itself. Part of it is the
"better" event handling to be used for testing.