dapp: add EthCC event prize for transfers
This features is only meant to be used for presentation on the EthCC
event to share prizes with the test users.
Instead of the normal blue check mark cycle, on a complete transfer, it
randomly shows either a beer or a t-shirt. This determines the prize the
tester gets. The randomize comes from the payment identifier which is
per default determined by a timestamp.
I decided to double the timeout for the pop-up to disappear. I made it
not 100% sticky, so the user still gets a quite normal experience.