sdk: add config.encryptSecret
since sending encrypted secrets use only target's pubkey, it doesn't
require our private key, so it can be enabled by default. target will of
course only be able to act on it if they've the private key available
(e.g. ether's Wallet, as default in CLI and with Raiden key)