dapp: downgrade Sass dependency
Unfortunately there is an issue of Vuetify which produces tons (massive amounts)
of warning output when building the dApp. This issue is known at the Vuetify
project. Unfortunately it can't be resolved while sticking to Vue 2 as it would
produce conflicting dependency versions. That is what I take away from the
respective issues (links below). This downgrade is the recommended solution.

https://github.com/vuetifyjs/vuetify/issues/13737
https://github.com/vuetifyjs/vuetify/issues/13694