dapp: set webpack performance configurations
Webpack prints warnings during the building of the dApp that some assets have
become to big for a nice performance. We are aware of this and are fine with it.
Especially because of the offline functionality. Therefore we must not see this
warning all the time.
I introduced a variable for a overall accepted maximum asset size. It gets also
used for the service worker caching configuration. If we should ever his this
limit, we have an actual issue.
If we want to improve performance, we can also use the lighthouse report tool.