dapp: simplify service worker registration
We have introduced an environment variable to disable the usage of the
service worker. Having this, we can use it as only flag to toggle the
feature. This means we must not check for production mode anymore. This
is also easily allows us to have the service worker in the end-to-end
tests.