dApp: improve Cypress test performance
This disables a code injection feature of Cypress. Cypress routes any
traffic through a proxy. This specific feature is known to cost a lot of
time. Just as a comparison: the dApp needs roughly 8s to cache all
assets in a normal web-browser. Within the Cypress test iframe it takes
over a minute. This makes it at least a little faster. The here usually
injected code is written to prevent the app to have some common issues
within an iframe. As we apparently don't suffer from these issues we can
disable it without any downside.

The second improvement disables the compression and storage of the test
video file for successful tests. This saves about one minute of
processing time for the best case. If there was an issue, it still takes
longer if it did not fail early.