dApp: fix page reload by new service workers
This is actually only necessary for a service worker that got activated
but took over the old local version. The control flow in the code was
wrong. This got discovered during the usage within the end-to-end tests
where the reload after the first service worker installation and
activation caused the tests to fail.